+++ src/app/pages/dashboard/points/points.component.ts
@@ -7,9 +7,11 @@ import { User } from '@interfaces/user'
 import { FirestoreService } from '@services/firestore.service'
 import { take } from 'rxjs/operators'
 import { DashboardService } from '../dashboard.service'
-import { DialogConfirmationComponent } from '../dialog-confirmation/dialog-confirmation.component'
-import { MatDialog, MatDialogRef } from '@angular/material'
-import { FormControl, Validators } from '@angular/forms'
+
+// import { DialogConfirmationComponent } from '../dialog-confirmation/dialog-confirmation.component'
+// import { MatDialog, MatDialogRef } from '@angular/material'
+import { FormControl } from '@angular/forms'
+import { ToastService } from '@services/toast.service'
 
 @Component({
   selector: 'app-points',
@@ -17,20 +19,19 @@ import { FormControl, Validators } from '@angular/forms'
   styleUrls: ['./points.component.scss']
 })
 export class PointsComponent {
-  dialogRef: MatDialogRef<DialogConfirmationComponent>
-  pointsInputForm = new FormControl('', Validators.required)
+  // dialogRef: MatDialogRef<DialogConfirmationComponent>
+  pointsInputForm = new FormControl('')
 
   constructor(
     public fss: FirestoreService,
-    public dialog: MatDialog,
+    // public dialog: MatDialog,
+    public toast: ToastService,
     private dashboardService: DashboardService,
     private afs: AngularFirestore
   ) {}
 
   // Add Points to one user onChange - Update the user value at the DB
-  update(user: User, newPoints: number, confirmed?: boolean) {
-    console.log('CONFIRMED', confirmed)
-
+  update(user: User, newPoints: number) {
     const userDoc: AngularFirestoreDocument<User> = this.afs.doc(
       `users/${user.uid}`
     )
@@ -39,27 +40,24 @@ export class PointsComponent {
 
     userData$.subscribe(userRef => {
       const addedPoints = newPoints - userRef.points
-      const points = confirmed ? newPoints : userRef.points
-      console.log('POINTS', points)
+      const points = newPoints
 
-      this.dashboardService.updateData(userDoc, { points })
-      if (confirmed) {
-        this.dashboardService.logData(points, addedPoints, userRef)
-      } else {
-        this.pointsInputForm.updateValueAndValidity()
-        // this.pointsInputForm.reset(this.pointsInputForm);
-        // console.log('THIS.POINTSINPUTFORM', this.pointsInputForm)
+      if (addedPoints > 700) {
+        return this.toast.showError(`Please add less than 700 points`)
       }
+
+      const newData: User = {
+        points
+      }
+
+      this.dashboardService.updateData(userDoc, newData)
+      this.dashboardService.logData(points, addedPoints, userRef)
     })
   }
 
-  openDialogConfirmation(user: User, newPoints: number): void {
-    this.dialogRef = this.dialog.open(DialogConfirmationComponent, {
-      autoFocus: false
-    })
-
-    this.dialogRef.beforeClose().subscribe((confirmed: boolean) => {
-      this.update(user, newPoints, confirmed)
-    })
-  }
+  // openDialogConfirmation(): void {
+  //   this.dialogRef = this.dialog.open(DialogConfirmationComponent, {
+  //     autoFocus: false
+  //   })
+  // }
 }